- Trabajo en clase con arquitectura SEP 22-23 Agosto.

+++ sep-sajja-datos/src/test/java/test/mx/sep/sajja/dao/UsuarioDaoTest.java
@@ -55,6 +55,6 @@ public class UsuarioDaoTest extends BaseDaoTest {
 	public void contarTest(){
 		Integer cuenta = usuarioDao.contar();
 		
-		Assert.assertEquals(new Integer(3), cuenta);
+		Assert.assertEquals(new Integer(4), cuenta);
 	}
 }
